servers/wolframalpha/requirements.txt
fastmcp
fastapi
uvicorn[standard]
python-dotenv
wolframalpha>=5.0.0 
httpx